requirements.txt
discord.py==2.3.2
googletrans==4.0.0-rc1